WindowFilterSettingsDialog: Replaced pykdeuic4 generated code with cleaned up ui file. Merged widget with dialog class and moved everything to windowfiltersettings.py. Ported all KDE elements to Qt.
