+++ lib/autokey/qtui/resources/resources.qrc
@@ -3,6 +3,7 @@
       <file alias="abbrsettings.ui">ui/abbrsettings.ui</file>
       <file alias="hotkeysettings.ui">ui/hotkeysettings.ui</file>
       <file alias="detectdialog.ui">ui/detectdialog.ui</file>
+      <file alias="window_filter_settings_dialog.ui">ui/window_filter_settings_dialog.ui</file>
   </qresource>
 </RCC>
