+++ lib/autokey/qtui/windowfiltersettings.py
@@ -1,59 +1,98 @@
-#!/usr/bin/env python
-# coding=UTF-8
-#
-# Generated by pykdeuic4 from windowfiltersettings.ui on Sat Mar 10 15:51:01 2012
-#
-# WARNING! All changes to this file will be lost.
-from PyKDE4 import kdecore
-from PyKDE4 import kdeui
-from PyQt4 import QtCore, QtGui
+# Copyright (C) 2011 Chris Dekter
+# Copyright (C) 2018 Thomas Hess <thomas.hess@udo.edu>
 
-try:
-    _fromUtf8 = QtCore.QString.fromUtf8
-except AttributeError:
-    _fromUtf8 = lambda s: s
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
 
-class Ui_Form(object):
-    def setupUi(self, Form):
-        Form.setObjectName(_fromUtf8("Form"))
-        Form.resize(425, 120)
-        self.verticalLayout = QtGui.QVBoxLayout(Form)
-        self.verticalLayout.setObjectName(_fromUtf8("verticalLayout"))
-        self.horizontalLayout = QtGui.QHBoxLayout()
-        self.horizontalLayout.setObjectName(_fromUtf8("horizontalLayout"))
-        self.label_2 = QtGui.QLabel(Form)
-        self.label_2.setObjectName(_fromUtf8("label_2"))
-        self.horizontalLayout.addWidget(self.label_2)
-        self.triggerRegexLineEdit = KLineEdit(Form)
-        self.triggerRegexLineEdit.setUrlDropsEnabled(False)
-        self.triggerRegexLineEdit.setProperty("showClearButton", True)
-        self.triggerRegexLineEdit.setObjectName(_fromUtf8("triggerRegexLineEdit"))
-        self.horizontalLayout.addWidget(self.triggerRegexLineEdit)
-        self.verticalLayout.addLayout(self.horizontalLayout)
-        self.recursiveCheckBox = QtGui.QCheckBox(Form)
-        self.recursiveCheckBox.setObjectName(_fromUtf8("recursiveCheckBox"))
-        self.verticalLayout.addWidget(self.recursiveCheckBox)
-        self.horizontalLayout_2 = QtGui.QHBoxLayout()
-        self.horizontalLayout_2.setObjectName(_fromUtf8("horizontalLayout_2"))
-        spacerItem = QtGui.QSpacerItem(40, 20, QtGui.QSizePolicy.Expanding, QtGui.QSizePolicy.Minimum)
-        self.horizontalLayout_2.addItem(spacerItem)
-        self.detectButton = QtGui.QPushButton(Form)
-        self.detectButton.setObjectName(_fromUtf8("detectButton"))
-        self.horizontalLayout_2.addWidget(self.detectButton)
-        self.verticalLayout.addLayout(self.horizontalLayout_2)
-        self.kseparator = KSeparator(Form)
-        self.kseparator.setObjectName(_fromUtf8("kseparator"))
-        self.verticalLayout.addWidget(self.kseparator)
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
 
-        self.retranslateUi(Form)
-        QtCore.QMetaObject.connectSlotsByName(Form)
+# You should have received a copy of the GNU General Public License
+# along with this program. If not, see <http://www.gnu.org/licenses/>.
 
-    def retranslateUi(self, Form):
-        Form.setWindowTitle(kdecore.i18n(_fromUtf8("Form")))
-        self.label_2.setText(kdecore.i18n(_fromUtf8("Regular expression to match:")))
-        self.triggerRegexLineEdit.setToolTip(kdecore.i18n(_fromUtf8("Window title")))
-        self.triggerRegexLineEdit.setWhatsThis(kdecore.i18n(_fromUtf8("Enter a regular expression that matches the title of windows in which you want this item to trigger.")))
-        self.recursiveCheckBox.setText(kdecore.i18n(_fromUtf8("Apply recursively to subfolders and items")))
-        self.detectButton.setText(kdecore.i18n(_fromUtf8("Detect Window Properties")))
+from PyQt4.QtGui import QDialog
+from PyQt4.QtGui import QFontMetrics
+from PyQt4.QtGui import QApplication
 
-from PyKDE4.kdeui import KSeparator, KLineEdit
+from .common import inherits_from_ui_file_with_name
+from .detectdialog import DetectDialog
+
+from .. import iomediator
+from .. import model
+
+
+# TODO: Once the port to Qt5 is done, enable the clearButtonEnable property for the line edit in the UI editor.
+# TODO: Pure Qt4 does not support the line edit clear button, so this functionality is currently unavailable.
+class WindowFilterSettingsDialog(*inherits_from_ui_file_with_name("window_filter_settings_dialog")):
+
+    def __init__(self, parent):
+        super(WindowFilterSettingsDialog, self).__init__(parent)
+        self.setupUi(self)
+        m = QFontMetrics(QApplication.font())
+        self.triggerRegexLineEdit.setMinimumWidth(m.width("windowclass.WindowClass"))
+        self.target_item = None
+        self.grabber = None  # type: iomediator.WindowGrabber
+
+    def load(self, item):
+        self.target_item = item
+
+        if not isinstance(item, model.Folder):
+            self.recursiveCheckBox.hide()
+        else:
+            self.recursiveCheckBox.show()
+
+        if not item.has_filter():
+            self.reset()
+        else:
+            self.triggerRegexLineEdit.setText(item.get_filter_regex())
+            self.recursiveCheckBox.setChecked(item.isRecursive)
+
+    def save(self, item):
+        item.set_window_titles(self.get_filter_text())
+        item.set_filter_recursive(self.get_is_recursive())
+
+    def get_is_recursive(self):
+        return self.recursiveCheckBox.isChecked()
+
+    def reset(self):
+        self.triggerRegexLineEdit.clear()
+        self.recursiveCheckBox.setChecked(False)
+
+    def reset_focus(self):
+        self.triggerRegexLineEdit.setFocus()
+
+    def get_filter_text(self):
+        return str(self.triggerRegexLineEdit.text())
+
+    def receive_window_info(self, info):
+        self.parentWidget().topLevelWidget().app.exec_in_main(self._receiveWindowInfo, info)
+
+    def _receiveWindowInfo(self, info):
+        dlg = DetectDialog(self)
+        dlg.populate(info)
+        dlg.exec_()
+
+        if dlg.result() == QDialog.Accepted:
+            self.triggerRegexLineEdit.setText(dlg.get_choice())
+
+        self.detectButton.setEnabled(True)
+
+    # --- Signal handlers ---
+
+    def on_detectButton_pressed(self):
+        self.detectButton.setEnabled(False)
+        self.grabber = iomediator.WindowGrabber(self)
+        self.grabber.start()
+
+    # --- event handlers ---
+
+    def slotButtonClicked(self, button):
+        if button == QDialog.Cancel:
+            self.load(self.targetItem)
+
+        QDialog.slotButtonClicked(self, button)
+
